requirements.txt
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
Pygments==2.9.0
pytz==2021.1
python-dotenv==0.19.0